requirements.txt
asgiref==3.8.1
Django==4.2.14
django-cors-headers==4.3.1
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
sqlparse==0.5.0
pip~=21.1.2
wheel~=0.36.2
setuptools~=57.0.0